+++ tests/test_client/test_KeystoneClient.py
